+++ src/common/buffer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#import "../bpf_helpers.h"
+#include "bpf_helpers.h"
 
 // maximum size of each buffer (since they may have flexible arrays at
 // the end)
