+++ src/common/common.h
