+++ src/common/offsets.h
