+++ src/process-events.c
@@ -6,11 +6,11 @@
 #include <linux/sched.h>
 #include <linux/uio.h>
 #include <linux/fcntl.h>
-#include "bpf_helpers.h"
-#include "types.h"
-#include "offsets.h"
-#include "repeat.h"
-#include "common.h"
+#include "common/bpf_helpers.h"
+#include "common/types.h"
+#include "common/offsets.h"
+#include "common/repeat.h"
+#include "common/common.h"
 
 #include "process/clone.h"
 #include "process/exec.h"
