+++ src/process/exec.h
@@ -1,9 +1,10 @@
 #pragma once
 
-#include "bpf_tracing.h"
-#include "buffer.h"
-#include "helpers.h"
-#include "path.h"
+#include "../common/bpf_tracing.h"
+#include "../common/buffer.h"
+#include "../common/helpers.h"
+#include "../common/path.h"
+
 #include "push_message.h"
 #include "script.h"
