+++ src/process/unshare.h
@@ -1,7 +1,9 @@
 #pragma once
 
-#include "../common.h"
-#include "helpers.h"
+#include "../common/common.h"
+#include "../common/helpers.h"
+
+#include "push_message.h"
 
 typedef struct {
   u64 pid_tgid;
